requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
certifi==14.05.14
gnureadline==6.3.3
ipython==2.2.0
lazy==1.2
matplotlib==1.4.0
mock==1.0.1
nose==1.3.4
numpy==1.9.0
pandas==0.14.1
patsy==0.3.0
pyparsing==2.0.2
python-dateutil==2.2
pytz==2014.7
requests==2.4.1
scipy==0.14.0
six==1.8.0
statsmodels==0.5.0
wsgiref==0.1.2
